chore: added  monokai graph theme
